+++ src/main/java/team/jit/technicalinterviewdemo/business/category/Category.java
@@ -1,6 +1,4 @@
 package team.jit.technicalinterviewdemo.business.category;
-
-import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -17,17 +15,14 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Table(name = "categories")
-@Schema(description = "Book category.")
 public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Setter(AccessLevel.NONE)
-    @Schema(description = "Database identifier.", example = "1")
     private Long id;
 
     @Column(nullable = false, unique = true)
-    @Schema(description = "Unique category name.", example = "Java")
     private String name;
 
     public Category(String name) {