+++ src/test/java/team/jit/technicalinterviewdemo/technical/api/ApiExceptionHandlerTests.java
@@ -30,7 +30,7 @@ class ApiExceptionHandlerTests {
 
     @BeforeEach
     void setUp() {
-        apiExceptionHandler = new ApiExceptionHandler(localizationMessageService);
+        apiExceptionHandler = new ApiExceptionHandler(new ApiProblemFactory(localizationMessageService));
     }
 
     @Test